Switch PRNG from xorshift7 to alea
xorshift7 has a flaw in its initialization that ignores the first
7 characters of the seed (see https://github.com/davidbau/seedrandom/issues/84)

The `seedrandom` package, while heavily used, does not appear to be
very rigorously tested after reviewing the source; consider switching
the package in the future. For now, alea seems to respect even single
character seeds and that's probably all we need.

Thanks to Twitch users inks000 for identifying the problem and
hexdefined for hunting down the details