Reduce double-dipping
The previous rephrasing of the bingo squares for clarity and consistency
had a problem: by phrasing things as succinct statements of truth
("Has X"), once-per-run events could be argued as valid multiple times
over long runs. The actual intent is that these events represent
_transitions_: X _becomes_ true.

Changed the wording and added clarifying notes for many of these
events. For example, it's no longer "Has 500+ Maximum Health"
(a statement which is true and remains true for the rest of the
run), but is now "Max Health Passes 500" (a statement which is
true exactly once, when the event happens, but not after)